Make get-output create a memoization context
This was missing from get-output. Previously, if there was a lot of
computation in a getter for an output wire (as opposed to, e.g.,
fetching a couple fields from the state struct), it would not benefit
from memoization of internal nodes in the circuit, which for even
moderately-sized examples makes a huge performance difference.