Fix bug in handling of persistent state in crash
Previously, the `crash+power-on-reset` function havoced state, stepped
the circuit for a single cycle with the reset line asserted, and then
copied over the persistent state. This is an incorrect order of
operations, because circuits can read from this persistent state in the
cycle where reset is asserted. This patch fixes the order of operations.